[react] Update React to disallow object child type
Objects are not valid as a React child. If you pass one, you'll receive
a runtime error "Objects are not valid as a React child", but the
component will compile without error. As an example, see the following
code sample:

``` ts
const nestedObject = { foo: { bar: 1 } };

const Component: React.SFC<{}> = () => <div>{nestedObject}</div>
```

Digging in, this seemed to be related to the `ReactFragment` type
definition which includes `{}`. Seemingly this is included to allow for
"render props", aka the function as child pattern.

With this change, I've replaced the use of `{}` in the `ReactFragment`
type definition with a newly defined `ReactChildFunction` definition to
ensure render props will still compile correctly, but a bare object
passed as a child will cause a compiler error.